+++ pkg/image/definition_test.go
@@ -42,13 +42,16 @@ func TestParse(t *testing.T) {
 	require.Len(t, userConfigs, 3)
 	assert.Equal(t, "alpha", userConfigs[0].Username)
 	assert.Equal(t, "$6$bZfTI3Wj05fdxQcB$W1HJQTKw/MaGTCwK75ic9putEquJvYO7vMnDBVAfuAMFW58/79abky4mx9.8znK0UZwSKng9dVosnYQR1toH71", userConfigs[0].EncryptedPassword)
-	assert.Contains(t, userConfigs[0].SSHKey, "ssh-rsa AAAAB3")
+	assert.Len(t, userConfigs[0].SSHKeys, 2)
+	assert.Contains(t, userConfigs[0].SSHKeys[0], "ssh-rsa AAAAB3")
+	assert.Contains(t, userConfigs[0].SSHKeys[1], "ssh-rsa BBBBB3")
 	assert.Equal(t, "beta", userConfigs[1].Username)
 	assert.Equal(t, "$6$GHjiVHm2AT.Qxznz$1CwDuEBM1546E/sVE1Gn1y4JoGzW58wrckyx3jj2QnphFmceS6b/qFtkjw1cp7LSJNW1OcLe/EeIxDDHqZU6o1", userConfigs[1].EncryptedPassword)
-	assert.Equal(t, "", userConfigs[1].SSHKey)
+	assert.Nil(t, userConfigs[1].SSHKeys)
 	assert.Equal(t, "gamma", userConfigs[2].Username)
 	assert.Equal(t, "", userConfigs[2].EncryptedPassword)
-	assert.Contains(t, userConfigs[2].SSHKey, "ssh-rsa BBBBB3")
+	assert.Len(t, userConfigs[2].SSHKeys, 1)
+	assert.Contains(t, userConfigs[2].SSHKeys[0], "ssh-rsa BBBBB3")
 
 	// Operating System -> Systemd
 	systemd := definition.OperatingSystem.Systemd