+++ RELEASE_NOTES.md
@@ -3,4 +3,7 @@
 ## Next
 
 * Added support for deploying user-provided Helm charts
+
+### Image Definition
 * Removed the `embeddedArtifactRegistry/images/supplyChainKey` attribute
+* Changed `operatingSystem/users/sshKey` into `operatingSystem/users/sshKeys` and it is now a list instead of a single string